requirements.txt
alembic==0.9.9
bcrypt==3.1.4
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
dokuwiki==1.1.0
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
libsass
lxml
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.8.4
pycparser==2.18
python-dateutil==2.7.2
python-editor==1.0.3
requests==2.18.4
rcssmin
six==1.12.0
SQLAlchemy==1.2.6
Unidecode==1.0.22
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
